+++ src/index.ts
@@ -98,11 +98,9 @@ class Vditor extends VditorMethod {
             this.vditor.upload = new Upload();
         }
 
-        // let lutePath = `http://192.168.2.248:9090/lute.min.js?${new Date().getTime()}`;
-        let lutePath = "src/js/lute/lute.min.js";
-        if (!options.debugger) {
-            lutePath = `${mergedOptions.cdn}/dist/js/lute/lute.min.js`;
-        }
+        // const lutePath = `http://192.168.2.248:9090/lute.min.js?${new Date().getTime()}`;
+        // const lutePath = "src/js/lute/lute.min.js";
+        const lutePath = `${mergedOptions.cdn}/dist/js/lute/lute.min.js`;
         addScript(lutePath, "vditorLuteScript").then(() => {
             this.vditor.lute = setLute({
                 autoSpace: this.vditor.options.preview.markdown.autoSpace,