+++ src/ts/ui/setCodeTheme.ts
@@ -1,7 +1,8 @@
-import {Constants} from "../constants";
+import {Constants, VDITOR_VERSION} from "../constants";
 import {addStyle} from "../util/addStyle";
 
-export const setCodeTheme = (codeTheme: string, cdn: string) => {
+export const setCodeTheme = (codeTheme: string,
+                             cdn = `https://cdn.jsdelivr.net/npm/vditor@${VDITOR_VERSION}`) => {
     if (!Constants.CODE_THEME.includes(codeTheme)) {
         codeTheme = "github";
     }
@@ -11,5 +12,6 @@ export const setCodeTheme = (codeTheme: string, cdn: string) => {
         addStyle(href, "vditorHljsStyle");
     } else if (vditorHljsStyle.href !== href) {
         vditorHljsStyle.remove();
+        addStyle(href, "vditorHljsStyle");
     }
 };